[scsi] Enable CONFIG_SCSI_LOGGING
Summary: CONFIG_SCSI_LOGGING was previously disabled during a FB kernel
config audit ~6 years ago.  This disabled/removed a bunch of config
values that were deemed to not be necessary D2698791.  We now need this
value enabled at least on test kernels as we have scsi issues  that need
it for debugging.

Test Plan: ```
$ ./facebook/scripts/prepareconfig x86_64 && make olddefconfig && make
-j$(nproc ) tar-pkg
$ kernelctl tardeploy linux-5.6.13-04658-gbbbefcf1cf72-x86.tar
rtptest1408.ftw5

$ sush rtptest1408.ftw5

$ dmesg | grep sd_done
```

Reviewers:

Subscribers:

Tasks: T78249002

Tags: top-of-stack

Differential Revision: https://phabricator.intern.facebook.com/D26404059