btrfs: don't fall back to buffered for weird alignments
Summary:
btrfs: don't fall back to buffered for weird alignments

Jens and Dan noticed in testing that 5.6 allows weird alignments with
O_DIRECT because we just fail back to buffered.  Upstream we don't do
this, we appropriately return an error.  This is preferred because then
the application doesn't assume it's allowed to get away with dumb
things.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Josef Bacik <jbacik@fb.com>