fb: mm: track page allocations statistics
Collect data about a number of desired and failed allocations
by order and expose it via the new /proc/page_allocs interface.

Use per-cpu counters to make the perf impact as low as possible.
Let's keep them racy (don't disable preemption etc), it doesn't
matter that much, we're looking for a high-level picture.

The output format is:
 <order> <allocation requests> <allocation failures>
 ...

Example output (after trying to allocate 10k 2MB hugepages):
  $ cat /proc/page_allocs
  0     3870608          0
  1        9610          0
  2        1992          0
  3        1382          0
  4          84          0
  5           9          0
  6           6          0
  7           3          0
  8           1          0
  9        2256          1
  10          3          0

NOTE: as now it's a fb-only patch. My plan is to gather some data
to have a better idea what's useful and what's not and also have
a better arguments towards upstream merging. Please, do not treat
this interface as stable, it can be changed or dismissed in the future.

Signed-off-by: Roman Gushchin <guro@fb.com>