kernel: don't signal io_uring owning task on worker exit
Since the io_uring io-wq workers are now effectively just implicitly
created threads from the parent, we also go through the usual parent
notification if one of these threads exit. As the task didn't ask for
the creation of these threads explicitly, don't send a SIGCHLD to the
parent when one of them exits.

Fixes: 3bfe6106693b ("io-wq: fork worker threads from original task")
Reported-by: Stefan Metzmacher <metze@samba.org>
Signed-off-by: Jens Axboe <axboe@kernel.dk>