mfd: intel-m10-bmc: Simplify the legacy version reg definition
The version register is the only one in the legacy I/O space to be
accessed, so it is not necessary to define the legacy base & version
register offset. A direct definition of the legacy version register
address would be fine.

Signed-off-by: Xu Yilun <yilun.xu@intel.com>
Reviewed-by: Tom Rix <trix@redhat.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
(cherry picked from commit 5893f4d1f43036664010e3ae1d3f7a98b2165a5d)
Signed-off-by: Jonathan Lemon <jonathan.lemon@gmail.com>