btrfs: don't free the reloc root if it's marked DEAD
Summary:
btrfs: don't free the reloc root if it's marked DEAD

If we get an error during relocation, we'll just drop the reloc root
without clearing root->reloc_root.  However this isn't great, because
later on we'll come along and unmount and free the reloc root again, and
then bad things happen.  There's a simple way to check for this,
BTRFS_ROOT_DEAD_RELOC_TREE being set means that we had this very case
happen, and we need to not trust root->reloc_root.  Handle this case in
the root freeing code by not doing our free if DEAD_RELOC_TREE is set.
This is fixed properly upstream with root ref counting, but that's too
invasive to backport.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Josef Bacik <jbacik@fb.com>