btrfs: take the lock for the bg before updating used
Summary:
btrfs: take the lock for the bg before updating used

We need to make sure the bg lock is held when updating used.  Not
strictly necessary for new block groups and on mount, but we have the
lockdep assert there for a reason, so take the lock.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Josef Bacik <jbacik@fb.com>